refactor(generator): 重构网站生成器代码并移除独立观察模块
